Config    -reverting to -O3
Field     -added field_descr to Block constructor
          -added field_descr to FieldBlock constructor
          -BUG FIX: FieldBlock::pup(): missing field_descr_ (broke checkpoint-1)
          -removed field_descr from FieldBlock::[re]allocate_array() parameters

Parameters -removed debug printf's
           -BUG FIX: double Parameters::value(<double>) return value--AGAIN
              (fixed before!)

Simulation  -BUG FIX: removed projections pup() (broke checkpoint-1)

+++ src/Cello/parameters_Parameters.hpp
@@ -67,7 +67,7 @@ public: // interface
   int value (std::string s, int deflt) throw()
   { return value_integer(s,deflt); }
 
-  int value (std::string s, double deflt) throw()
+  double value (std::string s, double deflt) throw()
   { return value_float(s,deflt); }
 
   int value (std::string s, bool deflt) throw()
@@ -79,7 +79,7 @@ public: // interface
   int value (int i,std::string s, int deflt) throw()
   { return list_value_integer(i,s,deflt); }
 
-  int value (int i,std::string s, double deflt) throw()
+  double value (int i,std::string s, double deflt) throw()
   { return list_value_float(i,s,deflt); }
 
   int value (int i,std::string s, bool deflt) throw()
